# backend/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Lobster",
    "description": "Go to /login/ to obtain your token. Authorize with the following value: 'Token {your_token}'"
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": [

      ]
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization"
      }
    }
  },
  "paths": {
...
    "/projects/{projectid}/files/{filename}/": {
      "get": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectid",
            "schema": {
              "type": "integer"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "filename",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "summary": "Gets a file from a project on lobster",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectid",
            "schema": {
              "type": "integer"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "filename",
            "schema": {
              "type": "string"
            },
            "required": true
          }
        ],
        "summary": "Gets a file from a project on lobster",
        "requestBody": {
          "description": "Exercise Post Object",
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/exercises/{exerciseid}/projects/": {
      "get": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "exerciseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Gets projects associated with an exercise on lobster",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "exerciseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "summary": "Adds a project to an exercise on lobster"
      }
    },
    "/projects/{projectid}/": {
      "get": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Gets details on a project on lobster",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "exerciseid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Deletes a project on lobster",
        "responses": {
          "204": {
            "description": "OK"
          }
        }
      }
    },
    "/projects/{projectid}/files/": {
      "get": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Gets starter_files for an exercise on lobster",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/projects/{projectid}/status/": {
      "get": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Gets the status on a current project",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Projects"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "projectid",
            "schema": {
              "type": "integer"
            },
            "required": true
          }
        ],
        "summary": "Updates the status on a current project",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "status": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          }
        }
      }
    },
...
  }
}